dev-requirements.txt
-r requirements.txt
djlint==1.36.4
Faker==35.2.0
pre_commit==4.1.0
pylint==3.3.4
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
ruff==0.9.5
